feat: revert changes on dashboard v2 metrics
