+++ apps/dapp/src/components/Pages/Core/DappPages/Dashboard/Table/TxnDataTable.tsx
@@ -1,8 +1,8 @@
 import { LoadingText } from 'components/Pages/Core/components/LoaderVault/commons/LoadingText';
 import env from 'constants/env/local';
-import { StrategyKey } from 'hooks/api/use-react-query';
 import styled from 'styled-components';
 import { loading } from 'utils/loading-value';
+import { StrategyKey } from '../hooks/use-dashboardv2-metrics';
 
 export type TableRow = {
   date: string;
