+++ apps/dapp/src/components/Pages/Core/DappPages/Dashboard/Table/TxnPaginationControl.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 type PaginationControlProps = {
   currentPage: number;
